# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elite FF Tournament Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestStorageUploadUrl
      tags: [storage]
      summary: Request a protected image upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                size:
                  type: number
                contentType:
                  type: string
              required: [name, size, contentType]
      responses:
        "200":
          description: Signed upload URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadURL:
                    type: string
                  objectPath:
                    type: string
                  metadata:
                    type: object
                required: [uploadURL, objectPath, metadata]
        "400":
          description: Invalid image metadata
        "401":
          description: Unauthorized
...
  /results:
    get:
      operationId: getResults
      tags: [results]
      summary: Get match results
      parameters:
        - name: tournamentId
          in: query
          schema:
            type: integer
        - name: registrationId
          in: query
          schema:
            type: integer
        - name: mine
          in: query
          schema:
            type: boolean
          description: Filter to only the current user's registrations
      responses:
        "200":
          description: Results list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchResult"
    post:
      operationId: createResult
      tags: [results]
      summary: Upload match result (Host only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResultBody"
      responses:
        "201":
          description: Result created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
  /results/{id}:
    put:
      operationId: updateResult
      tags: [results]
      summary: Update a result (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResultBody"
      responses:
        "200":
          description: Result updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
...
components:
  schemas:
...
    MatchResult:
      type: object
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        registrationId:
          type: integer
        matchNumber:
          type: integer
        squadName:
          type: string
        placement:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        kills:
          type: ["integer", "null"]
        prize:
          type: ["integer", "null"]
        prizeType:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        paymentScreenshotUrl:
          type: ["string", "null"]
        utrNumber:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - tournamentId
        - registrationId
        - matchNumber
        - squadName
        - createdAt
    CreateResultBody:
      type: object
      properties:
        tournamentId:
          type: integer
        registrationId:
          type: integer
        matchNumber:
          type: integer
        squadName:
          type: string
        placement:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        kills:
          type: ["integer", "null"]
        prize:
          type: ["integer", "null"]
        prizeType:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        paymentScreenshotUrl:
          type: ["string", "null"]
        utrNumber:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required:
        - tournamentId
        - registrationId
        - matchNumber
        - squadName
    UpdateResultBody:
      type: object
      properties:
        placement:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        kills:
          type: ["integer", "null"]
        prize:
          type: ["integer", "null"]
        prizeType:
          type: ["string", "null"]
        paymentScreenshotUrl:
          type: ["string", "null"]
        utrNumber:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...